# scheduler: internal data structure optimization