@Huang-Wei 2019-08-13
/hold cancel

Address updated and commits (except for the benchmark one) have been squashed.

@alculquicondor PTAL.