@Huang-Wei 2019-08-20
@ahg-g @alculquicondor Commits squashed. Please help to relabel /lgtm.

Thanks for the efforts reviewing this!